Don't try to find the closest clock, just use highest one.
  Previous patch wasn't fully correct, just that the values were generating
the same register dump as the smi 2.2.5 driver version.
  When correcting it to not mix Mhz and Khz in the math, the problems
returned.... Anyway, it just works when selecting the highest clock, and
the attached CRT will display the panel image. This is illogical, so maybe
the specs are outdated... The previous code has been #ifdef'ed out in
"CALC_CLOCK".
  Note that it still changes the boot default clock, from the default
288Mhz to 336Mhz, and in this patch, it also always sets both, panel and
crt clocks (what happens here, is that most likely, by doing this, it
forces the hardware to "somehow" autoconfigure itself; will try to get
some information from SMI about this, but this is the boot default,
and what smi sources do...).