Make Dualhead option functional.
  Slightly change clock selection code to start using 501 compatible values,
before checking 502 values, if it is a 502.
  DPMS for the VGA/second output was being set with bits inverted, that is
dpms-on was programmed as dpms-off and vice versa. This was one of the
reasons of dual head not working.
  Corrected wrong vsync programming for the crt. Cut&paste/typo caused
programming vsync with hsync values, and thus, never getting the crt
to accept the mode being programmed.
  If adding:
  Option "Dualhead" "True"
to xorg.conf, now you should be able to do things like:
$ DISPLAY=:0.0 xrandr --output VGA --right-of LVDS
$ DISPLAY=:0.0 xrandr --output VGA --below LVDS
and so on.
  *Iff* there is some way to not have any limitations for video playback,
this option should be made default or automatically configured.