Remove the ZoomOnLCD option and update manpage for PCIRetry.
  ZoomOnLCD controlled if Ctrl/Alt/+ or Ctrl/Alt/- could be used to
reprogram the mode on the panel. The option was not used, and now, if
different resolutions could be used, it should be advertised by randr.

  PCIBurst and PCIRetry options are now enabled by default.

  The fifo_xxx options need a review, as they appear to be associated
with the older WaitQueue macro, that did ignore its argument. But these
options are not used, and may be removed or reimplemented at some time.